Reorganize tests into separate folder and add snapshot testing
